requirements.txt
annotated-types==0.7.0
attrs==24.3.0
cachetools==5.5.0
cycler==0.12.1
datasets==3.2.0
distro==1.9.0
huggingface-hub==0.27.0
ipykernel==6.29.5
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multiprocess==0.70.12.2
numpy==1.26.4
paddlepaddle-gpu==3.0.0.dev20250123
pandas==2.2.3
pillow==10.4.0
PyYAML==6.0.2
scikit-learn==1.6.0
scipy==1.14.1
sympy==1.13.3
tqdm==4.67.1